Define the abstract syntax.
The abstract syntax pretty closely follows the grammar. A few points
worth noting:

1. Every expression is tagged with an optional type. At parse time,
these are all None. Then the Typecheck phase will attempt to fill
in the type annotations, alerting the user if it is unable to do so.

2. Every Absyn Element is tagged with a location (line and column)
in the input text which generated that element. These can be used
to provide useful error messages even after most of the concrete
syntax information has been lost, such as at typecheck time.