Add support for parameterized chains.
For instance, you can now write:
my_source(freq, amp, index) = {
  fm(freq, amp, index)
}

where freq, amp, and index evaluate to numbers.

To greatly simplify the typecheck phase while still accomodating
this increased complexity, we treat all assignments as functions,
with possibly zero arguments. For instance, at least until
typechecking is complete, the assignment
foo = 42
will be treated as a 0-ary function
foo() = 42
At the moment, we do not support parameterized functions aside from
chains, so
foo(param) = param
is ill-formed. However, the infrastructure is in place to change
this fairly easily. That change will become more useful in the
future when arithmetic expressions are added to the language.