Implement a translator from CSound++ abstract syntax to CSound.
The translator implements all features currently supported by the
parser and typechecker, EXCEPT:

Channel maps are not yet implemented. Instead, each instrument is
mapped to the first unmapped MIDI channel, in the order in which
instruments are defined.

The translations produced by the translator also require a CSound
preamble in order to run correctly. This preamble is not yet written,
but will include such things as function table definitions and built-in
CSPP functions (such as fm, compress, etc).