Implement a simple import system.
Before the main part of a program, the user can have any number
of import clauses. The line
import "path/to/file.csp"
will cause the indicated file to be parsed and the definitions
therein substituted into the importing file. Files are looked up
first as relative or absolute paths, as appropriate. If that fails,
the system attempts to find a file by interpreting the path as
relative to resources. This is where the standard library will be
kept.