Create a CSound code minifier.
In an effort to keep the size of generated files small, we remove
all blank lines, comments, and unnecessary whitespace before we
output any CSound code.

Note that a failure in the minifier is not fatal to the compilation
process; we'll still output the longer file if we have to.

This utility is also useful for testing, when we want to compare
generated code to some expectation modulo everything the minifier
clears out. This is basically a cheap substitute for parsing the
CSound code and comparing it to an abstract syntax tree.