Abstract multiplexers, sources, and compressors with a component object.
Components can have an arbitrary number of inputs and outputs (not both
zero). In a chain, the output-arity of one component must equal the input-
arity of the next. The input-arity of the chain is that of the first element,
and the output-arity is that of the last element.

In a parallel block, the input-arity is the sum of the input-arities of the
components, and the output-arity is the sum of the output-arities.

For both chains and parallel blocks, an empty block is implicitly translated
to a one-input, one-output block that leaves its input unchanged.

The concept formerly known as Source is now simply a component with no inputs
and one output. Likewise, an Effect is a component with one input and one
output.

An instrument is still required to be a source.