Redesign typechecker to allow parameter overloading and reordering.
Symbols are now looked up by a (name, arity) tuple, instead of
looking up a name and asserting on the arity. This allows multiple
symbols to be defined with the same name so long as they have
different arities.

The system also reorders the statements in a program, via the
following method:
1. Separate assignments from instruments
2. Construct a (name, arity) -> assignment map
3. For each assignment in the map
    3a. If it has any dependencies (applications) typecheck them
    3b. Typecheck the symbol, remove it from the map, and add it to
        the back of a list of checked symbols.
4. Typecheck each instrument.
This procedure ensures that any given symbol is translated after all
of its dependencies, and thus programmers can define symbols in any
order they like.

Recursion is still not allowed, and there is a mechanism for detecting
recursion of any cycle length and reporting it to the user as an error.