Refactor typechecker to encapsulate state in a monad.
The previous idiom for updating the context:
  (context2, resultA) = someFunction(context)
  (context3, resultB) = someOtherFunction(context2, resultA)
was tedious, ugly, and highly error prone. This is now replaced by
a monad which has an internal context. The context is automatically
updated and passed forward via the bind operation (flatMap), and is
accessed and updated via a specialized api. This allows us to write
code like the following:
  for {
    resultA <- someFunction
    resultB <- someOtherFunction(resultA)
  }
where the context is passed forward implicitly.

This cleans up the code by eliminating the need for recursive helper
functions whenever we need to process a list, and it completely
eliminates the class of errors that arises from forgetting to change
the number of a context, or changing it somewhere and forgetting to
change it elsewhere.