Factor primitives out of typechecker.
They are now in a dedicated object SwPrimitives which contains,
along with the names, arities, and types of the primitives,
definitions which allow some of them (at present, only numeric
functions) to be evaluated at compile time. This will allow for
compile-time evaluation as an optimization as well as a tool for
verifying correctness. For example, it will soon be desirable to
evaluate the channel expressions in an instrument definition at
compile time, so that we can check for dependencies between
instruments.